fix: exchange download button
