+++ src/containers/sidebars/national-report-sidebar/national-report-sidebar.js
@@ -5,14 +5,15 @@ import Component from './national-report-sidebar-component';
 import * as urlActions from 'actions/url-actions';
 import metadataActions from 'redux_modules/metadata';
 import countryDataActions from 'redux_modules/country-data';
+import { downloadNrcPdfAnalytics } from 'actions/google-analytics-actions';
+
 import { NATIONAL_REPORT_CARD_LANDING, NATIONAL_REPORT_CARD } from 'router'
 
-const actions = { ...urlActions, ...countryDataActions, ...metadataActions };
+const actions = { ...urlActions, ...countryDataActions, ...metadataActions, downloadNrcPdfAnalytics };
 
 const NationalReportSidebarContainer = (props) => {
   const {
     browsePage,
-    handlePrintReport,
     onboardingType,
   } = props;
 
@@ -29,6 +30,16 @@ const NationalReportSidebarContainer = (props) => {
       changeUI({ onboardingStep: onboardingStep + 1, onboardingType, waitingInteraction: false });
     }
   };
+  const handlePrintReport = () => {
+    const { downloadNrcPdfAnalytics, countryName } = props;
+    const today = new Date();
+    const date = Intl.DateTimeFormat('en-US', { month: 'long', year: 'numeric' }).format(today);
+    const tempTitle = document.title;
+    document.title = `Half-Earth National Report Card ${date} - ${countryName}`;
+    window.print();
+    downloadNrcPdfAnalytics(countryName)
+    document.title = tempTitle;
+  }
 
   return (
     <Component
