+++ src/containers/sidebars/national-report-sidebar/overview-sidebar/index.js
@@ -2,33 +2,21 @@ import React, { useEffect } from 'react';
 import Component from './component.jsx';
 import { connect } from 'react-redux';
 import mapStateToProps from 'containers/sidebars/national-report-sidebar/national-report-sidebar-selectors';
-import { visitNrcOverviewAnalytics, downloadNrcPdfAnalytics } from 'actions/google-analytics-actions';
-const actions = { visitNrcOverviewAnalytics, downloadNrcPdfAnalytics }
+import { visitNrcOverviewAnalytics } from 'actions/google-analytics-actions';
+const actions = { visitNrcOverviewAnalytics }
 
 const OverviewSidebar = (props) => {
 
-    useEffect(() => {
-      const { visitNrcOverviewAnalytics } = props;
-      visitNrcOverviewAnalytics()
-    }, [])
+  useEffect(() => {
+    const { visitNrcOverviewAnalytics } = props;
+    visitNrcOverviewAnalytics()
+  }, [])
 
-    const handlePrintReport = () => {
-      const { downloadNrcPdfAnalytics, countryName } = props;
-      const today = new Date();
-      const date = Intl.DateTimeFormat('en-US', { month: 'long', year: 'numeric'}).format(today);
-      const tempTitle = document.title;
-      document.title = `Half-Earth National Report Card ${date} - ${countryName}`;
-      window.print();
-      downloadNrcPdfAnalytics(countryName)
-      document.title = tempTitle;
-    }
-
-    return (
-      <Component
-        handlePrintReport={handlePrintReport}
-        {...props}
-      />
-    )
+  return (
+    <Component
+      {...props}
+    />
+  )
 }
 
 export default connect(mapStateToProps, actions)(OverviewSidebar);
