--- src/containers/sidebars/national-report-sidebar/overview-sidebar/styles.module.scss
@@ -1,7 +1,0 @@
-@import 'styles/settings';
-
-.actionButton {
-    margin-left: $site-gutter;
-    margin-bottom: $site-gutter;
-  }
-
